+++ packages/user-management-widget/src/lib/index.ts
@@ -1,0 +1,16 @@
+import { UserManagementWidget } from './widget';
+
+declare global {
+  interface HTMLElement {
+    attributeChangedCallback(
+      attrName: string,
+      oldValue: string | null,
+      newValue: string | null,
+    ): void;
+    connectedCallback(): void;
+  }
+}
+
+export default UserManagementWidget;
+
+customElements.define('descope-user-management-widget', UserManagementWidget);