+++ packages/user-management-widget/src/lib/widget/api/sdk/createUserSdk.ts
@@ -1,0 +1,150 @@
+import {
+  CreateUserConfig,
+  HttpClient,
+  SearchUsersConfig,
+  User,
+} from '../types';
+import { apiPaths } from '../apiPaths';
+import { withErrorHandler } from './helpers';
+
+export const createUserSdk = ({
+  httpClient,
+  tenant,
+}: {
+  httpClient: HttpClient;
+  tenant: string;
+}) => {
+  const search: (config: SearchUsersConfig) => Promise<User[]> = async ({
+    page,
+    limit = 10000,
+    customAttributes,
+    statuses,
+    emails,
+    phones,
+    text,
+    sort,
+  } = {}) => {
+    const res = await httpClient.post(
+      apiPaths.user.search,
+      {
+        limit,
+        page,
+        withTestUser: false,
+        customAttributes,
+        statuses,
+        emails,
+        phones,
+        text,
+        sort,
+      },
+      {
+        queryParams: { tenant },
+      },
+    );
+
+    await withErrorHandler(res);
+
+    const json = await res.json();
+
+    return json.users;
+  };
+
+  const deleteBatch = async (userIds: string[]) => {
+    const res = await httpClient.post(
+      apiPaths.user.deleteBatch,
+      { userIds },
+      {
+        queryParams: { tenant },
+      },
+    );
+
+    await withErrorHandler(res);
+
+    return res.json();
+  };
+
+  const create: (config: CreateUserConfig) => Promise<User[]> = async ({
+    loginId,
+    email,
+    phone,
+    displayName,
+    roles,
+    customAttributes,
+    picture,
+    verifiedEmail,
+    verifiedPhone,
+    givenName,
+    middleName,
+    familyName,
+    additionalLoginIds,
+    sendSMS,
+    sendMail,
+    inviteUrl,
+    invite,
+  }) => {
+    const res = await httpClient.post(
+      apiPaths.user.create,
+      {
+        loginId,
+        email,
+        phone,
+        displayName,
+        givenName,
+        middleName,
+        familyName,
+        roleNames: roles,
+        userTenants: [{ tenantId: tenant }],
+        customAttributes,
+        picture,
+        verifiedEmail,
+        verifiedPhone,
+        additionalLoginIds,
+        sendSMS,
+        sendMail,
+        inviteUrl,
+        invite,
+      },
+      {
+        queryParams: { tenant },
+      },
+    );
+
+    await withErrorHandler(res);
+
+    const json = await res.json();
+
+    return json.user;
+  };
+
+  const expirePassword = async (loginIds: string[]) => {
+    const res = await httpClient.post(
+      apiPaths.user.expirePassword,
+      { loginId: loginIds[0] },
+      {
+        queryParams: { tenant },
+      },
+    );
+
+    await withErrorHandler(res);
+
+    return res.json();
+  };
+
+  const getCustomAttributes = async () => {
+    const res = await httpClient.get(apiPaths.user.customAttributes, {
+      queryParams: { tenant },
+    });
+
+    await withErrorHandler(res);
+
+    return res.json();
+  };
+
+  return {
+    search,
+    deleteBatch,
+    create,
+    expirePassword,
+    getCustomAttributes,
+  };
+};